# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  models.Product:
    properties:
      category:
        example: gi
        type: string
      created_at:
        type: string
      description:
        example: Premium BJJ gi with excellent fit and durability
        type: string
      id:
        example: 1
        type: integer
      image_url:
        example: https://example.com/gi.jpg
        type: string
      name:
        example: Tatami Estilo 6.0 Gi
        type: string
      price:
        example: 120
        type: number
      size_options:
        description: Changed to simple string
        example: A1,A2,A3,A4
        type: string
      stock:
        example: 15
        type: integer
      updated_at:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: admin@bjjstore.com
    name: Calvin Le
    url: http://www.bjjstore.com/support
  description: A complete e-commerce API for Brazilian Jiu-Jitsu gear and equipment
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: BJJ Store API
  version: "1.0"
paths:
...
  /admin/products:
    post:
      consumes:
        - application/json
      description: Create a new product (Admin only)
      parameters:
        - description: Product data
          in: body
          name: product
          required: true
          schema:
            $ref: "#/definitions/models.Product"
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/models.Product"
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Insufficient permissions
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
        - BearerAuth: []
      summary: Create a new product
      tags:
        - admin
        - products
  /admin/products/{id}:
    delete:
      consumes:
        - application/json
      description: Delete a product by ID (Admin only)
      parameters:
        - description: Product ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid product ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Insufficient permissions
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
        - BearerAuth: []
      summary: Delete a product
      tags:
        - admin
        - products
    put:
      consumes:
        - application/json
      description: Update an existing product (Admin only)
      parameters:
        - description: Product ID
          in: path
          name: id
          required: true
          type: integer
        - description: Updated product data
          in: body
          name: product
          required: true
          schema:
            $ref: "#/definitions/models.Product"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Product"
        "400":
          description: Invalid request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Insufficient permissions
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
        - BearerAuth: []
      summary: Update a product
      tags:
        - admin
        - products
...
  /products:
    get:
      consumes:
        - application/json
      description: Get a list of all products with optional filtering
      parameters:
        - description: Filter by category
          in: query
          name: category
          type: string
        - description: Search in product name and description
          in: query
          name: search
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/models.Product"
            type: array
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all products
      tags:
        - products
  /products/{id}:
    get:
      consumes:
        - application/json
      description: Get a single product by its ID
      parameters:
        - description: Product ID
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/models.Product"
        "400":
          description: Invalid product ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Product not found
          schema:
            additionalProperties: true
            type: object
      summary: Get a product by ID
      tags:
        - products
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"